Use isLoading for better routing + add Loading component
