+++ src/App.tsx
@@ -5,23 +5,24 @@ import Dashboard from './Components/Dashboard';
 import Header from './Components/Header';
 import Login from './Components/Login';
 import HomeDesktop from './Components/HomeDesktop';
-import LoginCallback from './Components/LoginCallback';
+import Loading from './Components/Loading';
 import { useWalletMode } from './Hooks/useWallet';
 import './App.css';
 
 function App() {
-  const { claims, logout } = useCriiptoVerify();
+  const { claims, logout, isLoading, isInitializing } = useCriiptoVerify();
   let mql = window.matchMedia('(min-width: 1024px)');
   const walletMode = useWalletMode();
+  const getRedirectPath = (basePath: string) => (walletMode ? `${basePath}?wallet=true` : basePath);
 
   const handleLogout = () => {
-    logout({ redirectUri: walletMode ? `${window.location.origin}/?wallet=true` : window.location.origin });
+    logout({redirectUri: getRedirectPath(window.location.origin)});
   };
 
   const loginRoutes = (
     <>
       <Route path="/login" element={<Login />} />
-      <Route path="/login/callback" element={<LoginCallback />} />
+      <Route path="/login/callback" element={<Navigate to={getRedirectPath('/login')} />} />
       <Route path="/dashboard" element={<Navigate to="/login" />} />
     </>
   );
@@ -31,34 +32,33 @@ function App() {
       <Header handleLogout={handleLogout} claims={claims} />
 
       <Routes>
-        {claims ? (
+        {isLoading || isInitializing ? (
+          <Route index element={<Loading />} />
+        ) : claims ? (
           <>
             <Route index element={<Navigate to="/dashboard" />} />
-            <Route path="/login/callback" element={<LoginCallback />} />
-            <Route path="/login" element={<Navigate to="/dashboard" />} />
+            <Route path="/login/callback" element={<Navigate to={getRedirectPath('/dashboard')} />}/>
+            <Route path="/login" element={<Navigate to={getRedirectPath('/dashboard')} />} />
             <Route path="/dashboard" element={<Dashboard claims={claims} />} />
-          </>
-        ) : mql.matches ? (
-          <>
-            <Route index element={<HomeDesktop />} />
-            {loginRoutes}
           </>
         ) : (
           <>
-            <Route index element={<Navigate to="/login" />} />
+            <Route index element={mql.matches ? <HomeDesktop /> : <Navigate to={getRedirectPath('/login')} />}/>
             {loginRoutes}
           </>
         )}
-        <Route
-          path="*"
-          element={
-            <div className="flex justify-center align-center">
-              <h1 className="font-semibold text-2xl m-16 ml-4 mt-16 leading-normal">
-                Page not found
-              </h1>
-            </div>
-          }
-        />
+        {!isLoading && !isInitializing && !claims && (
+          <Route
+            path="*"
+            element={
+              <div className="flex justify-center align-center">
+                <h1 className="font-semibold text-2xl m-16 ml-4 mt-16 leading-normal">
+                  Page not found
+                </h1>
+              </div>
+            }
+          />
+        )}
       </Routes>
     </div>
   );
